feat: single page editor support select attachments
